refactor: 불필요하게 마지막 프레임에 인덱스를 넘기는 정적 팩토리 생성자 수정
